加入了详细的日志

+++ source/settings.h
@@ -6,7 +6,7 @@ class Settings
 private:
     QFile* settingsFile;
     QFile* logFile;
-    short m_size;
+    double m_size;
     short m_timerInterval;
     bool m_doAutoAlign;
 public:
@@ -24,7 +24,7 @@ public:
     //写日志
     void write_log(const QString& text, LogType tp = LogType::info);
 
-    short size() {return m_size;}
+    double size() {return m_size;}
     short timerInterval() {return m_timerInterval;}
     bool doAutoAlign() {return m_doAutoAlign;}
 };
